+++ alerting-config-service-impl/src/main/java/org/hypertrace/alerting/config/service/AlertingConfigRequestValidator.java
@@ -1,5 +1,6 @@
 package org.hypertrace.alerting.config.service;
 
+import com.google.common.base.Preconditions;
 import org.hypertrace.alerting.config.service.v1.CreateNotificationChannelRequest;
 import org.hypertrace.alerting.config.service.v1.CreateNotificationRuleRequest;
 import org.hypertrace.alerting.config.service.v1.DeleteNotificationChannelRequest;
@@ -11,35 +12,105 @@ import org.hypertrace.alerting.config.service.v1.UpdateNotificationChannelReques
 import org.hypertrace.alerting.config.service.v1.UpdateNotificationRuleRequest;
 import org.hypertrace.core.grpcutils.context.RequestContext;
 
-public interface AlertingConfigRequestValidator {
-  void validateCreateNotificationRuleRequest(
+public class AlertingConfigRequestValidator {
+  
+  public void validateCreateNotificationRuleRequest(
       RequestContext requestContext,
       CreateNotificationRuleRequest request,
-      AlertingConfigServiceCoordinator alertingConfigServiceCoordinator);
+      NotificationConfigServiceStore notificationConfigServiceStore) {
+    validateTenantID(requestContext);
+    Preconditions.checkArgument(
+        request.hasNewNotificationRule(), "Notification rule should be present");
+    Preconditions.checkArgument(
+        !request.getNewNotificationRule().getRuleName().isBlank(), "Rule name cannot be empty");
+    Preconditions.checkArgument(
+        !request.getNewNotificationRule().getChannelId().isBlank(), "ChannelId cannot be empty");
+    notificationConfigServiceStore.getNotificationChannel(
+        requestContext, request.getNewNotificationRule().getChannelId());
+  }
 
-  void validateUpdateNotificationRuleRequest(
+  
+  public void validateUpdateNotificationRuleRequest(
       RequestContext requestContext,
       UpdateNotificationRuleRequest request,
-      AlertingConfigServiceCoordinator alertingConfigServiceCoordinator);
+      NotificationConfigServiceStore  notificationConfigServiceStore) {
+    validateTenantID(requestContext);
+    Preconditions.checkArgument(
+        request.hasNotificationRule(), "Notification rule should be present");
+    Preconditions.checkArgument(
+        !request.getNotificationRule().getRuleName().isBlank(), "Rule name cannot be empty");
+    Preconditions.checkArgument(
+        !request.getNotificationRule().getChannelId().isBlank(), "ChannelId cannot be empty");
+    notificationConfigServiceStore.getNotificationChannel(
+        requestContext, request.getNotificationRule().getChannelId());
+  }
 
-  void validateGetAllNotificationRulesRequest(
-      RequestContext requestContext, GetAllNotificationRulesRequest request);
+  
+  public void validateGetAllNotificationRulesRequest(
+      RequestContext requestContext, GetAllNotificationRulesRequest request) {
+    validateTenantID(requestContext);
+  }
 
-  void validateDeleteNotificationRuleRequest(
-      RequestContext requestContext, DeleteNotificationRuleRequest request);
+  
+  public void validateDeleteNotificationRuleRequest(
+      RequestContext requestContext, DeleteNotificationRuleRequest request) {
+    validateTenantID(requestContext);
+    Preconditions.checkArgument(
+        !request.getNotificationRuleId().isBlank(), "RuleId cannot be empty");
+  }
 
-  void validateCreateNotificationChannelRequest(
-      RequestContext requestContext, CreateNotificationChannelRequest request);
+  
+  public void validateCreateNotificationChannelRequest(
+      RequestContext requestContext, CreateNotificationChannelRequest request) {
+    validateTenantID(requestContext);
+    Preconditions.checkArgument(
+        request.hasNewNotificationChannel(), "Notification channel should be present");
+    Preconditions.checkArgument(
+        !request.getNewNotificationChannel().getChannelName().isBlank(),
+        "Channel name should be present");
+    Preconditions.checkArgument(
+        request.getNewNotificationChannel().hasNotificationChannelConfig(),
+        "Channel config has to be present");
+  }
 
-  void validateUpdateNotificationChannelRequest(
-      RequestContext requestContext, UpdateNotificationChannelRequest request);
+  
+  public void validateUpdateNotificationChannelRequest(
+      RequestContext requestContext, UpdateNotificationChannelRequest request) {
+    validateTenantID(requestContext);
+    Preconditions.checkArgument(
+        request.hasNotificationChannel(), "Notification channel should be present");
+    Preconditions.checkArgument(
+        !request.getNotificationChannel().getChannelName().isBlank(),
+        "Channel name should be present");
+    Preconditions.checkArgument(
+        request.getNotificationChannel().hasNotificationChannelConfig(),
+        "Channel config has to be present");
+  }
 
-  void validateGetAllNotificationChannelsRequest(
-      RequestContext requestContext, GetAllNotificationChannelsRequest request);
+  
+  public void validateGetAllNotificationChannelsRequest(
+      RequestContext requestContext, GetAllNotificationChannelsRequest request) {
+    validateTenantID(requestContext);
+  }
 
-  void validateGetNotificationChannelRequest(
-      RequestContext requestContext, GetNotificationChannelRequest request);
+  
+  public void validateGetNotificationChannelRequest(
+      RequestContext requestContext, GetNotificationChannelRequest request) {
+    validateTenantID(requestContext);
+    Preconditions.checkArgument(
+        !request.getNotificationChannelId().isBlank(), "ChannelId cannot be empty");
+  }
 
-  void validateDeleteNotificationChannelRequest(
-      RequestContext requestContext, DeleteNotificationChannelRequest request);
+  
+  public void validateDeleteNotificationChannelRequest(
+      RequestContext requestContext, DeleteNotificationChannelRequest request) {
+    validateTenantID(requestContext);
+    Preconditions.checkArgument(
+        !request.getNotificationChannelId().isBlank(), "ChannelId cannot be empty");
+  }
+
+  private void validateTenantID(RequestContext requestContext) {
+    Preconditions.checkArgument(
+        requestContext.getTenantId().isPresent(), "Tenant ID is not present in request context");
+  }
 }
